implementacionflaskprueba-main/src/app.py
from flask import Flask
from routes.contacts import contacts
from decouple import config
from flask_marshmallow import Marshmallow
from flask_cors import CORS

app = Flask(__name__)
CORS(app)

db_name = config('DB_NAME')
db_user = config('DB_USER',)
db_password = config('DB_PASSWORD')
db_host = config('DB_HOST')
db_port = config('DB_PORT')

app.config["SQLALCHEMY_DATABASE_URI"] = f"postgresql+psycopg2://{db_user}:{db_password}@{db_host}:{db_port}/{db_name}"
app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
app.config['SEND_FILE_MAX_AGE_DEFAULT'] = 0
Marshmallow(app)

app.register_blueprint(contacts, url_prefix='/api/contacts')