implementacionflaskprueba-main/src/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.3
flask-marshmallow==1.2.1
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
mysqlclient==2.2.4
packaging==24.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing-extensions==4.11.0
werkzeug==3.0.2
zipp==3.18.1
